Fix usage before assignment
When new connection is created in doGetAsyncDedicatedConnection(), it is
not yet assigned to asyncDedicatedConnection attribute, so it cannot be
used in potentiallySelectDatabase() call.

It seems that the select call is superfluous anyway here, because the
`select()` method call can be made only when the cached dedicated
connection is used (not a shared one) and the `select()` method always
selects the database immediately on the cached dedicated connection.

Therefore it should never be necessary to call select on newly created
dedicated connection, because it is either created by `select()` method
call or a default database is used.

Closes #2984

Signed-off-by: Oldřich Jedlička <oldium.pro@gmail.com>